Refactor: Moved sampling and renamed some generators.
